pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hcdp-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for Hawai'i Climate Data Portal API"
authors = [
    {name = "HCDP MCP", email = "hcdp@hawaii.edu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
hcdp-mcp-server = "hcdp_mcp_server.server:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hcdp_mcp_server*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"